# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Demande un code OTP par SMS",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Vérifie l'OTP et émet les tokens JWT",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rotation du refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Révoque le refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Utilisateur courant",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/devotionals/today": {
      "get": {
        "operationId": "DevotionalsController_today",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dévotion du jour",
        "tags": [
          "devotionals"
        ]
      }
    },
    "/api/v1/devotionals": {
      "get": {
        "operationId": "DevotionalsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historique des dévotions",
        "tags": [
          "devotionals"
        ]
      },
      "post": {
        "operationId": "DevotionalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDevotionalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée une dévotion (staff)",
        "tags": [
          "devotionals"
        ]
      }
    },
    "/api/v1/devotionals/{id}": {
      "get": {
        "operationId": "DevotionalsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail dévotion",
        "tags": [
          "devotionals"
        ]
      },
      "patch": {
        "operationId": "DevotionalsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDevotionalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour / planifie (staff)",
        "tags": [
          "devotionals"
        ]
      },
      "delete": {
        "operationId": "DevotionalsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Supprime un brouillon (staff)",
        "tags": [
          "devotionals"
        ]
      }
    },
    "/api/v1/devotionals/{id}/read": {
      "post": {
        "operationId": "DevotionalsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marque lu + met à jour le streak",
        "tags": [
          "devotionals"
        ]
      }
    },
...
    "/api/v1/events": {
      "get": {
        "operationId": "EventsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des événements",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "EventsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée un événement (staff)",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}": {
      "get": {
        "operationId": "EventsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail événement",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "EventsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour un événement (staff)",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{id}/rsvp": {
      "post": {
        "operationId": "EventsController_rsvp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RsvpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "RSVP à un événement",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/appointments/topics": {
      "get": {
        "operationId": "AppointmentsController_topics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Sujets de rendez-vous",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/slots": {
      "get": {
        "operationId": "AppointmentsController_slots",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Créneaux disponibles",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/mine": {
      "get": {
        "operationId": "AppointmentsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mes rendez-vous",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments": {
      "get": {
        "operationId": "AppointmentsController_agenda",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agenda staff",
        "tags": [
          "appointments"
        ]
      },
      "post": {
        "operationId": "AppointmentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Prend un rendez-vous",
        "tags": [
          "appointments"
        ]
      }
    },
    "/api/v1/appointments/{id}": {
      "delete": {
        "operationId": "AppointmentsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Annule un rendez-vous",
        "tags": [
          "appointments"
        ]
      },
      "patch": {
        "operationId": "AppointmentsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppointmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour un rendez-vous (staff)",
        "tags": [
          "appointments"
        ]
      }
    },
...
  },
  "info": {
    "title": "IPCK API",
    "description": "API de la plateforme IPCK (mobile + dashboard)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+243810000000",
            "description": "Numéro au format E.164"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+243810000000"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "CreateDevotionalDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "verseRef": {
            "type": "string"
          },
          "verseText": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "prayer": {
            "type": "string"
          },
          "applyTitle": {
            "type": "string"
          },
          "applySteps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published"
            ]
          },
          "publishAt": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        },
        "required": [
          "date",
          "title",
          "verseRef",
          "verseText",
          "body",
          "prayer",
          "applyTitle",
          "applySteps"
        ]
      },
      "UpdateDevotionalDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "verseRef": {
            "type": "string"
          },
          "verseText": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "prayer": {
            "type": "string"
          },
          "applyTitle": {
            "type": "string"
          },
          "applySteps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "scheduled",
              "published"
            ]
          },
          "publishAt": {
            "type": "string"
          },
          "author": {
            "type": "string"
          }
        }
      },
...
      "RsvpDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "going",
              "cancelled"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "CreateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "whenLabel": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "startsAt"
        ]
      },
      "UpdateEventDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startsAt": {
            "type": "string"
          },
          "whenLabel": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "CreateAppointmentDto": {
        "type": "object",
        "properties": {
          "topicId": {
            "type": "string",
            "example": "counseling"
          },
          "slotStart": {
            "type": "string",
            "example": "2026-06-02T14:00:00.000Z"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "topicId",
          "slotStart"
        ]
      },
      "UpdateAppointmentDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "tentative",
              "confirmed",
              "cancelled"
            ]
          },
          "pastorId": {
            "type": "string"
          },
          "slotStart": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        }
      },
...
    }
  }
}